Feat/178 restructure pia (#223)
* Restructuring of attack file.

* Fix number of classes for meta classifier. #178
