+++ privacy_evaluator/attacks/__init__.py
@@ -3,4 +3,4 @@ Module providing attacks.
 """
 from . import membership_inference
 from .attack import Attack
-from .property_inference_attack import PropertyInferenceAttack
+from . import property_inference_attack
