+++ privacy_evaluator/attacks/property_inference_attack/__init__.py
@@ -1,0 +1,7 @@
+from .property_inference_attack import PropertyInferenceAttack
+from .property_inference_class_distribution_attack import (
+    PropertyInferenceClassDistributionAttack,
+)
+from .property_inference_data_augmentation_attack import (
+    PropertyInferenceDataAugmentationAttack,
+)
