+++ tests/attacks/property_inference_attack/property_inference_data_augmentation_test.py
@@ -1,4 +1,6 @@
-from privacy_evaluator.attacks.property_inference_attack import PropertyInferenceAttack
+from privacy_evaluator.attacks.property_inference_attack.property_inference_data_augmentation_attack import (
+    PropertyInferenceDataAugmentationAttack,
+)
 from privacy_evaluator.classifiers.classifier import Classifier
 from privacy_evaluator.utils.data_utils import (
     dataset_downloader,
@@ -9,7 +11,7 @@ from privacy_evaluator.output.user_output_property_inference_attack import (
 from privacy_evaluator.utils.model_utils import create_and_train_torch_ConvNet_model
 from privacy_evaluator.utils.data_utils import create_new_dataset_with_adaptation
 
-from typing import Dict, List
+from typing import List
 import numpy as np
 import logging
 
@@ -33,6 +35,8 @@ ADAPTATION = "mask"
 RATIO = 0.2
 # Involved when adaptation is "mask", the side length of masking boxes.
 BOX_LEN = 4
+# number of epochs for training the meta classifier
+NUM_EPOCHS_META_CLASSIFIER = 2
 
 
 def test_property_inference_data_augmentation_attack(
@@ -46,6 +50,7 @@ def test_property_inference_data_augmentation_attack(
     box_len: int = BOX_LEN,
     adaptation: str = ADAPTATION,
     ratio: float = RATIO,
+    num_epochs_meta_classifier: int = NUM_EPOCHS_META_CLASSIFIER,
 ):
     logger = logging.getLogger(__name__)
     if verbose == 2:
@@ -75,4 +80,41 @@ def test_property_inference_data_augmentation_attack(
         model, "sparse_categorical_crossentropy", num_classes, input_shape
     )
 
-    # TODO start attack
+    logger.info("Start attack ...")
+
+    attack = PropertyInferenceDataAugmentationAttack(
+        target_model,
+        train_dataset,
+        amount_sets=amount_sets,
+        size_set=size_set,
+        ratios_for_attack=ratios_for_attack,
+        verbose=verbose,
+        num_epochs_meta_classifier=num_epochs_meta_classifier,
+        adaptation=adaptation,
+        box_len=box_len,
+    )
+
+    output = attack.attack()
+    assert (
+        attack.input_shape == input_shape
+    ), f"Wrong input shape. Input shape should be {input_shape}."
+    assert (
+        attack.amount_sets >= 2 and attack.amount_sets % 2 == 0
+    ), "Number of shadow classifiers must be even and greater than 1."
+    # we expect the ratios to be ordered
+    ratios_for_attack.sort()
+
+    assert isinstance(
+        output, UserOutputPropertyInferenceAttack
+    ), "Wrong output type of attack."
+    assert (
+        attack.ratios_for_attack == ratios_for_attack
+    ), "Ratios for properties are not equal to input."
+    assert (
+        attack.amount_sets == amount_sets
+    ), "Number of shadow classifiers are not equal to input."
+    assert attack.size_set == size_set, "Number of samples is not equal to input."
+    assert attack.adaptation == adaptation, "Classes are not equal to input classes."
+    assert len(output.output) == len(
+        ratios_for_attack
+    ), "Output is not compatible to input."
